[#7 fix] genetic Algorithm V4 - 수면 시간의 startTime이 24시간 전일 때의 오류 해결
